+++ src/daemon-command/logs.daemon-command.ts
@@ -13,15 +13,9 @@ export function logsDaemonCommand(daemon: Daemon, socket: Socket, names: string[
 
     for (const script of scriptsToProcess) {
         for (const line of script.logBuffer) {
-            socket.write(`${script.name}: ${line}\n`);
+            socket.write(`${script.logPrefix}${line}\n`);
         }
 
-        script.logSockets.push(socket);
-        socket.on("close", () => {
-            const index = script.logSockets.findIndex((i) => i == socket);
-            if (index !== -1) {
-                script.logSockets.splice(index, 1);
-            }
-        });
+        script.addLogSocket(socket);
     }
 }