+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 -   pm config file isn't configurable anymore
 -   waitOn support directly in script definition
 -   renamed binary to dev-pm
+-   added --follow option to start/restart command to follow logs after start
 -   add cpu/memory usage to status output
 -   add --interval option to status output to keep status open and optionally refresh periodically